+++ examples/noise_generator.rs
@@ -3,7 +3,9 @@
 use std::error::Error;
 
 fn main() -> Result<(), Box<dyn Error>> {
-    use rodio::source::{pink, white, Source};
+    use rodio::source::{
+        blue, brownian, gaussian_white, pink, triangular_white, velvet, violet, white, Source,
+    };
     use std::thread;
     use std::time::Duration;
 
@@ -19,10 +21,56 @@ fn main() -> Result<(), Box<dyn Error>> {
 
     thread::sleep(interval_duration);
 
+    stream_handle.mixer().add(
+        gaussian_white(48000)
+            .amplify(0.1)
+            .take_duration(noise_duration),
+    );
+    println!("Playing Gaussian white noise");
+
+    thread::sleep(interval_duration);
+
+    stream_handle.mixer().add(
+        triangular_white(48000)
+            .amplify(0.1)
+            .take_duration(noise_duration),
+    );
+    println!("Playing triangular white noise");
+
+    thread::sleep(interval_duration);
+
     stream_handle
         .mixer()
         .add(pink(48000).amplify(0.1).take_duration(noise_duration));
     println!("Playing pink noise");
+
+    thread::sleep(interval_duration);
+
+    stream_handle
+        .mixer()
+        .add(blue(48000).amplify(0.1).take_duration(noise_duration));
+    println!("Playing blue noise");
+
+    thread::sleep(interval_duration);
+
+    stream_handle
+        .mixer()
+        .add(violet(48000).amplify(0.1).take_duration(noise_duration));
+    println!("Playing violet noise");
+
+    thread::sleep(interval_duration);
+
+    stream_handle
+        .mixer()
+        .add(brownian(48000).amplify(0.1).take_duration(noise_duration));
+    println!("Playing brownian noise");
+
+    thread::sleep(interval_duration);
+
+    stream_handle
+        .mixer()
+        .add(velvet(48000).amplify(0.1).take_duration(noise_duration));
+    println!("Playing velvet noise");
 
     thread::sleep(interval_duration);